Update and rename ApexVod_UI.py to apex_vod_highlights.py
